Code simplification and refactor
