+++ WordPress/src/main/java/org/wordpress/android/support/he/model/VideoDownloadState.kt
@@ -5,6 +5,6 @@ import java.io.File
 sealed class VideoDownloadState {
     object Idle : VideoDownloadState()
     object Downloading : VideoDownloadState()
+    object Error : VideoDownloadState()
     data class Success(val file: File) : VideoDownloadState()
-    data class Error(val message: String) : VideoDownloadState()
 }
